refactor(schema): split engine runtime type modules
